Updated dashboard UI background
Linked new welcome screen to dashboard,
Updated dashboard screen background to match new theme
