+++ pubspec.yaml
@@ -14,7 +14,7 @@ description: A code generator for flutter
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.5.0 <3.0.0"
 
 dependencies:
   flutter:
